chore: renamed `document_details` to `document_upload`
