+++ frontend/app/src/main/java/com/example/thinkr/app/Route.kt
@@ -29,9 +29,9 @@ sealed interface Route {
     }
 
     @Serializable
-    data class DocumentDetails(val selectedUri: String) : Route {
+    data class DocumentUpload(val selectedUri: String) : Route {
         companion object {
-            const val ROUTE = "documentDetails/{selectedUri}"
+            const val ROUTE = "documentUpload/{selectedUri}"
             const val ARGUMENT = "selectedUri"
             fun createRoute(selectedUri: Uri): String {
                 return ROUTE.replace("{selectedUri}", Uri.encode(selectedUri.toString()))
