+++ frontend/app/src/main/java/com/example/thinkr/ui/document_upload/DocumentUploadState.kt
@@ -1,8 +1,8 @@
-package com.example.thinkr.ui.document_details
+package com.example.thinkr.ui.document_upload
 
 import android.net.Uri
 
-data class DocumentDetailsState(
+data class DocumentUploadState(
     val name: String = "document name",
     val context: String = "context",
     val uri: Uri = Uri.EMPTY
